Add improved `parents` type
Previously, a basic `Parent` from `@types/unist` was used, as an array,
for the second parameter of a visitor (`parents`).
This changes that to instead use an array of descendants in `tree`
which implement the abstract `Parent` interface.
This is not perfect, because several parents can’t be found in certain
nodes practically, but it will at least help folks narrow.

Related to syntax-tree/unist-util-visit#30.